fix(downloader): store sanitized url, store relative paths, correctly link creator
The url of the music-track wasn't stored properly, it is now sanitized.
Relative paths are now stored instead of full paths, this is more future
proof. The creator is now linked correctly to the music tracks. There
was a bug that if the creator was already existing it wouldn't link it
properly.